+++ core/src/main/groovy/org/hidetake/groovy/ssh/connection/Connection.groovy
@@ -3,10 +3,10 @@ package org.hidetake.groovy.ssh.connection
 import com.jcraft.jsch.*
 import groovy.util.logging.Slf4j
 import org.hidetake.groovy.ssh.core.Remote
-import org.hidetake.groovy.ssh.extension.settings.LocalPortForwardSettings
-import org.hidetake.groovy.ssh.extension.settings.RemotePortForwardSettings
 import org.hidetake.groovy.ssh.session.BackgroundCommandException
 import org.hidetake.groovy.ssh.session.BadExitStatusException
+import org.hidetake.groovy.ssh.session.forwarding.LocalPortForwardSettings
+import org.hidetake.groovy.ssh.session.forwarding.RemotePortForwardSettings
 
 /**
  * A SSH connection.
