--- core/src/main/groovy/org/hidetake/groovy/ssh/extension/CoreExtensions.groovy
@@ -1,17 +1,0 @@
-package org.hidetake.groovy.ssh.extension
-
-/**
- * A set of extensions to be shipped as default.
- *
- * @author Hidetake Iwata
- */
-trait CoreExtensions implements
-        Command,
-        BackgroundCommand,
-        Shell,
-        Sudo,
-        FileGet,
-        FilePut,
-        SftpRemove,
-        PortForward {
-}
