+++ core/src/main/groovy/org/hidetake/groovy/ssh/operation/DefaultOperations.groovy
@@ -3,8 +3,8 @@ package org.hidetake.groovy.ssh.operation
 import groovy.util.logging.Slf4j
 import org.hidetake.groovy.ssh.connection.Connection
 import org.hidetake.groovy.ssh.core.Remote
-import org.hidetake.groovy.ssh.extension.settings.LocalPortForwardSettings
-import org.hidetake.groovy.ssh.extension.settings.RemotePortForwardSettings
+import org.hidetake.groovy.ssh.session.forwarding.LocalPortForwardSettings
+import org.hidetake.groovy.ssh.session.forwarding.RemotePortForwardSettings
 
 import static org.hidetake.groovy.ssh.util.Utility.callWithDelegate
