+++ core/src/main/groovy/org/hidetake/groovy/ssh/operation/Operations.groovy
@@ -1,8 +1,8 @@
 package org.hidetake.groovy.ssh.operation
 
 import org.hidetake.groovy.ssh.core.Remote
-import org.hidetake.groovy.ssh.extension.settings.LocalPortForwardSettings
-import org.hidetake.groovy.ssh.extension.settings.RemotePortForwardSettings
+import org.hidetake.groovy.ssh.session.forwarding.LocalPortForwardSettings
+import org.hidetake.groovy.ssh.session.forwarding.RemotePortForwardSettings
 
 /**
  * An aggregate of core SSH operations.
